chore: moved networks.js file
